Accept host option
This change provides support for specifying the particular imgix host,
or source, that is to be used. Given that we are no longer relying on
the imgix source to be configured via the Rails initializer, we remove
the validate_configuration! method. Also, when we pass the options to
the to_url method we ensure that the host option is not passed.